Allow overriding the redis connection class
Add an additional optional parameter to the RedisChannelLayer
which accepts a django module import string. This allows overriding
the default redis.Redis connection client with a compatible alternate
implementation. This is specifically useful for running asgi_redis
against a redis cluster, as redis-py-cluster's implementation can be
used. For clusters, both prefix and stats_prefix should be surrounded
by {} (for example {asgi}:) to ensure that all channels related keys
are hashed into the same shard. Otherwise, various EVALSHA commands
will fail with messages about keys not hashing to the same slot.